Last of the examples to run.
Adds the Chapel and C version of the numpy example for nightly testing.
