--- docs/source/examples/chapel.numpy.synthetic.py
@@ -1,49 +1,0 @@
-#!/usr/bin/env python
-import argparse
-import numpy as np
-from pych.extern import Chapel
-
-@Chapel()
-def simulation(tsteps=int, spin=np.ndarray, decay=np.ndarray, velocity=np.ndarray):
-    """
-    var delta = velocity;
-    for timestep in 1..tsteps {
-        delta = (spin * decay + velocity) / delta;
-    }
-    return +reduce(delta);
-    """
-    return float
-
-def main(tsteps, particles):
-    """Load data, run simulation, visualize results."""
-
-    spin = np.ones(particles)               # Load data
-    decay = np.ones(particles)
-    velocity = np.ones(particles)
-
-    res = simulation(                       # Run simulation
-        tsteps,
-        spin,
-        decay,
-        velocity
-    )
-                                            # Visualize result
-    print("Phenomenon after %d tsteps = %d." % (tsteps, res))
-
-if __name__ == "__main__":                  # Argument parsing
-    parser = argparse.ArgumentParser(
-        description='Example illustrating simulation code.'
-    )
-    parser.add_argument(
-        '--tsteps', help="# of tsteps in 'simulation'",
-        type=int,
-        required=True
-    )
-    parser.add_argument(
-        '--particles', help="# of elements in arrays '",
-        type=int,
-        required=True
-    )
-    args = parser.parse_args()
-
-    main(args.tsteps, args.particles)
